Refactoring code: Creat a component Folder for NavBar
