+++ components/Components_TabBar/TabBarIcon.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Icon } from 'expo';
 
-import Colors from '../constants/Colors';
+import Colors from '../../constants/Colors';
 
 export default class TabBarIcon extends React.Component {
   render() {
